Reorganize serde modules
